+++ src/main/java/com/gildedgames/the_aether/entities/hostile/EntityUligo.java
@@ -52,7 +52,7 @@ public class EntityUligo extends EntityLiving implements IMob
         this.dataWatcher.addObject(16, (byte) 1);
         this.dataWatcher.addObject(15, (byte) -1);
         this.dataWatcher.addObject(17, (byte) 0);
-        this.dataWatcher.addObject(18, Byte.valueOf((byte)0));
+        this.dataWatcher.addObject(18, (byte) 0);
     }
     
     public int getMaxSafePointTries()
@@ -107,7 +107,7 @@ public class EntityUligo extends EntityLiving implements IMob
     public void onStruckByLightning(EntityLightningBolt p_70077_1_)
     {
         super.onStruckByLightning(p_70077_1_);
-        this.dataWatcher.updateObject(17, Byte.valueOf((byte)1));
+        this.dataWatcher.updateObject(17, (byte) 1);
     }
 
     /**
@@ -125,7 +125,7 @@ public class EntityUligo extends EntityLiving implements IMob
 
         this.setSlimeSize(i + 1);
         
-        this.dataWatcher.updateObject(17, Byte.valueOf((byte)(p_70037_1_.getBoolean("powered") ? 1 : 0)));
+        this.dataWatcher.updateObject(17, (byte) (p_70037_1_.getBoolean("powered") ? 1 : 0));
 
         if (p_70037_1_.hasKey("Fuse", 99))
         {
@@ -518,7 +518,7 @@ public class EntityUligo extends EntityLiving implements IMob
 
     public void func_146079_cb()
     {
-        this.dataWatcher.updateObject(18, Byte.valueOf((byte)1));
+        this.dataWatcher.updateObject(18, (byte) 1);
     }
     
     /**
@@ -534,7 +534,7 @@ public class EntityUligo extends EntityLiving implements IMob
      */
     public void setCreeperState(int p_70829_1_)
     {
-        this.dataWatcher.updateObject(15, Byte.valueOf((byte)p_70829_1_));
+        this.dataWatcher.updateObject(15, (byte) p_70829_1_);
     }
     
     public boolean canDespawn() {
